String Padding
Just call this method on a string variable. Provide "left" or "right" as the first argument for which side you want to pad. The second argument is the character you want to pad the string with. The final argument is the length you want the string to be after padding has been applied.